Move the namespace query param up the route hierarchy
This makes it active on all job routes
